+++ weblate/utils/requirements.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2012 - 2018 Michal Čihař <michal@cihar.com>
+# Copyright © 2012 - 2019 Michal Čihař <michal@cihar.com>
 #
 # This file is part of Weblate <https://weblate.org/>
 #
@@ -285,6 +285,13 @@ def get_versions():
         'https://github.com/jamesturk/jellyfish',
         'jellyfish',
         '0.6.1',
+    ))
+
+    result.append(get_single(
+        'diff-match-patch',
+        'https://github.com/diff-match-patch-python/diff-match-patch',
+        'diff_match_patch',
+        '20121119',
     ))
 
     return result